+++ packages/browser-wallet/src/popup/pages/NetworkSettings/i18n/en.ts
@@ -1,11 +1,6 @@
 const t = {
-    save: 'Save',
-    url: {
-        label: 'JSON-RPC endpoint',
-        validation: {
-            required: 'JSON-PRC endpoint is required',
-        },
-    },
+    connect: 'Connect',
+    connected: 'Connected',
 };
 
 export default t;