+++ packages/browser-wallet/src/popup/shell/Routes.tsx
@@ -8,6 +8,7 @@ import { absoluteRoutes, relativeRoutes } from '@popup/constants/routes';
 import MainLayout from '@popup/page-layouts/MainLayout';
 import FullscreenPromptLayout from '@popup/page-layouts/FullscreenPromptLayout';
 import Account from '@popup/pages/Account';
+import Identity from '@popup/pages/Identity';
 import SignMessage from '@popup/pages/SignMessage';
 import SendTransaction from '@popup/pages/SendTransaction';
 import Setup from '@popup/pages/Setup';
@@ -17,8 +18,10 @@ import Settings from '@popup/pages/Settings';
 import NetworkSettings from '@popup/pages/NetworkSettings';
 import VisualSettings from '@popup/pages/VisualSettings';
 import AddAccount from '@popup/pages/AddAccount';
+import { IdentityIssuanceEnd, IdentityIssuanceStart } from '@popup/pages/IdentityIssuance';
 import About from '@popup/pages/About';
 import Login from '@popup/pages/Login/Login';
+import ChangePasscode from '@popup/pages/ChangePasscode/ChangePasscode';
 
 type PromptKey = keyof Omit<typeof absoluteRoutes['prompt'], 'path'>;
 
@@ -48,10 +51,6 @@ function useMessagePrompt<R>(type: InternalMessageType | MessageType, promptKey:
     return handleResponse;
 }
 
-function NoContent() {
-    return <>No content yet...</>;
-}
-
 export default function Routes() {
     const handleConnectionResponse = useMessagePrompt<boolean>(InternalMessageType.Connect, 'connectionRequest');
     const handleSendTransactionResponse = useMessagePrompt<string | undefined>(
@@ -61,6 +60,11 @@ export default function Routes() {
     const handleSignMessageResponse = useMessagePrompt<AccountTransactionSignature | undefined>(
         InternalMessageType.SignMessage,
         'signMessage'
+    );
+
+    const handleIdentityIssuanceResponse = useMessagePrompt<void>(
+        InternalMessageType.EndIdentityIssuance,
+        'endIdentityIssuance'
     );
 
     useEffect(() => {
@@ -97,14 +101,22 @@ export default function Routes() {
                         />
                     }
                 />
+                <Route
+                    path={relativeRoutes.prompt.endIdentityIssuance.path}
+                    element={<IdentityIssuanceEnd onFinish={handleIdentityIssuanceResponse} />}
+                />
             </Route>
             <Route path={`${relativeRoutes.setup.path}/*`} element={<Setup />} />
             <Route path={relativeRoutes.login.path} element={<Login />} />
             <Route path={relativeRoutes.home.path} element={<MainLayout />}>
-                <Route element={<NoContent />} path={relativeRoutes.home.identities.path} />
+                <Route
+                    element={<IdentityIssuanceStart />}
+                    path={`${relativeRoutes.home.identities.path}/${relativeRoutes.home.identities.add.path}`}
+                />
+                <Route element={<Identity />} path={relativeRoutes.home.identities.path} />
                 <Route path={relativeRoutes.home.settings.path}>
                     <Route index element={<Settings />} />
-                    <Route element={<NoContent />} path={relativeRoutes.home.settings.passcode.path} />
+                    <Route element={<ChangePasscode />} path={relativeRoutes.home.settings.passcode.path} />
                     <Route element={<NetworkSettings />} path={relativeRoutes.home.settings.network.path} />
                     <Route element={<VisualSettings />} path={relativeRoutes.home.settings.visual.path} />
                     <Route element={<About />} path={relativeRoutes.home.settings.about.path} />