+++ packages/browser-wallet/src/popup/shell/i18n/locales/da.ts
@@ -10,8 +10,10 @@ import networkSettings from '@popup/pages/NetworkSettings/i18n/da';
 import visualSettings from '@popup/pages/VisualSettings/i18n/da';
 import about from '@popup/pages/About/i18n/da';
 import addAccount from '@popup/pages/AddAccount/i18n/da';
+import identityIssuance from '@popup/pages/IdentityIssuance/i18n/da';
 import login from '@popup/pages/Login/i18n/da';
 import transactionLog from '@popup/pages/Account/TransactionLog/i18n/da';
+import changePasscode from '@popup/pages/ChangePasscode/i18n/en';
 
 import type en from './en';
 
@@ -27,9 +29,11 @@ const t: typeof en = {
     networkSettings,
     visualSettings,
     addAccount,
+    identityIssuance,
     about,
     login,
     transactionLog,
+    changePasscode,
 };
 
 export default t;