+++ packages/browser-wallet/src/popup/shell/i18n/locales/en.ts
@@ -10,8 +10,10 @@ import networkSettings from '@popup/pages/NetworkSettings/i18n/en';
 import visualSettings from '@popup/pages/VisualSettings/i18n/en';
 import about from '@popup/pages/About/i18n/en';
 import addAccount from '@popup/pages/AddAccount/i18n/en';
+import identityIssuance from '@popup/pages/IdentityIssuance/i18n/en';
 import login from '@popup/pages/Login/i18n/en';
 import transactionLog from '@popup/pages/Account/TransactionLog/i18n/en';
+import changePasscode from '@popup/pages/ChangePasscode/i18n/en';
 
 const t = {
     shared,
@@ -25,9 +27,11 @@ const t = {
     networkSettings,
     visualSettings,
     addAccount,
+    identityIssuance,
     about,
     login,
     transactionLog,
+    changePasscode,
 };
 
 export default t;